rename directory to cat
