+++ mempool/cat/mempool.go
@@ -1,4 +1,4 @@
-package v2
+package cat
 
 import (
 	"errors"
