+++ mempool/cat/mempool_bench_test.go
@@ -1,4 +1,4 @@
-package v2
+package cat
 
 import (
 	"fmt"
