+++ mempool/cat/mempool_test.go
@@ -1,4 +1,4 @@
-package v2
+package cat
 
 import (
 	"bytes"
